requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
greenlet==1.1.2
h11==0.13.0
idna==3.3
outcome==1.2.0
playwright==1.24.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pyee==8.1.0
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.28.1
rsa==4.9
selenium==4.3.0
selenium-stealth==1.0.6
sniffio==1.2.0
sortedcontainers==2.4.0
Telethon==1.25.4
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
websockets==10.1
wsproto==1.1.0
